requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
CacheControl==0.14.3
certifi==2025.4.26
charset-normalizer==3.4.2
DateTime==5.5
distro==1.9.0
docker==7.1.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.3
openai==1.99.6
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
propcache==0.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.9.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
zope.interface==7.2
zstandard==0.23.0aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.11.0
CacheControl==0.14.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.3.1
DateTime==5.5
distro==1.9.0
docker==7.1.0
eval_type_backport==0.2.2
filelock==3.20.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.3.5
openai==1.99.6
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
propcache==0.3.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.2.0
setuptools==80.9.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
together==1.5.31
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
zope.interface==7.2
zstandard==0.23.0
